Added views for graphSystem
